pyproject.toml
[project]
name = "langchain-aig-response-tracing"
version = "1.0.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "click>=8.1.7",
    "langchain>=0.3.6",
    "langchain-openai>=0.2.5",
    "pangea-sdk>=5.2.0b2",
    "pydantic>=2.9.2",
]

[tool.uv]
dev-dependencies = [
    "mypy>=1.13.0",
    "ruff>=0.7.1",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    # ruff defaults
    "E4",
    "E7",
    "E9",
    "F",

    "I",   # isort
    "PTH", # flake8-use-pathlib
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]
